bot/requirements.txt
aiogram
aiohttp
sqlalchemy
python-dotenv